refactor(wallet): move watchingAlert into wallet domain
