+++ src/features/rnbw-rewards/screens/rnbw-rewards-screen/scenes/AirdropIntroScene.tsx
@@ -11,10 +11,10 @@ import {
 } from '@/features/rnbw-rewards/animations/sceneTransitions';
 import { RnbwRewardsScenes } from '@/features/rnbw-rewards/screens/rnbw-rewards-screen/constants/rewardsScenes';
 import { rewardsFlowActions } from '@/features/rnbw-rewards/stores/rewardsFlowStore';
+import { watchingAlert } from '@/features/wallet/utils/watchingAlert';
 import * as i18n from '@/languages';
 import { useIsReadOnlyWallet } from '@/state/wallets/walletsStore';
 import { time } from '@/utils/time';
-import watchingAlert from '@/utils/watchingAlert';
 
 const enteringAnimation = createScaleInFadeInSlideEnterAnimation({ delay: time.ms(200) });
