+++ src/screens/claimables/shared/components/ClaimButton.tsx
@@ -7,9 +7,9 @@ import ShimmerAnimation from '@/components/animations/ShimmerAnimation';
 import { HoldToActivateButton } from '@/components/hold-to-activate-button/HoldToActivateButton';
 import { enableActionsOnReadOnlyWallet } from '@/config/debug';
 import { AccentColorProvider, Box, Inline, Text, TextShadow } from '@/design-system';
+import { watchingAlert } from '@/features/wallet/utils/watchingAlert';
 import { getIsReadOnlyWallet } from '@/state/wallets/walletsStore';
 import deviceUtils from '@/utils/deviceUtils';
-import watchingAlert from '@/utils/watchingAlert';
 
 const BUTTON_WIDTH = deviceUtils.dimensions.width - 52;
