+++ src/screens/claimables/sponsored/components/SponsoredClaimableFlow.tsx
@@ -1,9 +1,9 @@
 import React, { useCallback, useMemo } from 'react';
 
+import { watchingAlert } from '@/features/wallet/utils/watchingAlert';
 import * as i18n from '@/languages';
 import { useNavigation } from '@/navigation/Navigation';
 import { useIsReadOnlyWallet } from '@/state/wallets/walletsStore';
-import watchingAlert from '@/utils/watchingAlert';
 
 import { ClaimButton } from '../../shared/components/ClaimButton';
 import { ClaimPanel } from '../../shared/components/ClaimPanel';
